Fixing user following errors
Previously, the following method was returning a post to the
users/follows endpoint, which doesn't exist. Changing it to return a get
to the same endpoint fixes this.
With that fix, the endpoint now returns a list of channels, but
UserFollowsResponse extended an arraylist of Users instead. Updating the
extended arraylist template fixes the returned data to match what it's
supposed to be.